aider: Changed the background of the webpage to file src/campus.jpg.
# Aider chat conversation:

USER: can you change the background of the webpage to file src/campus.jpg

ASSISTANT: To change the background of the webpage to an image, we need to modify the CSS file that controls the styling of the webpage. In this case, it's the `App.css` file. We will add a new rule to set the background image of the body to `src/campus.jpg`.

Here is the *edit block* for the change:

```css
workspaces/mySwap/src/App.css
<<<<<<< HEAD
.App {
  text-align: center;
  background-color: #100E17;
  color: #FFFFFF;
  font-family: 'Inter', sans-serif;
}
=======
.App {
  text-align: center;
  background-image: url('./campus.jpg');
  background-size: cover;
  color: #FFFFFF;
  font-family: 'Inter', sans-serif;
}
>>>>>>> updated
```

Please note that the image file `campus.jpg` should be located in the `src` directory for this to work. If the image is located elsewhere, you need to adjust the path accordingly.